% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fonction.R
\name{gen_game}
\alias{gen_game}
\title{gen_game}
\usage{
gen_game(text, open = FALSE)
}
\arguments{
\item{text}{text to use}

\item{open}{booleen open browser}
}
\description{
generate a game on ztype
}
\examples{
require(ztype)
require(magrittr)
c("dplyr","ggplot2","lubridate") \%>\% gen_liste() \%>\%
gen_paragraphe(10) \%>\%
gen_game() \%>\% browseURL()
}

