% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fonction.R
\name{gen_liste}
\alias{gen_liste}
\title{gen_liste}
\usage{
gen_liste(packages)
}
\arguments{
\item{packages}{pacakge to parse}
}
\description{
generate a list of function
}
\examples{
require(ztype)
require(magrittr)
c("dplyr","ggplot2","lubridate") \%>\% gen_liste()
}

