% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fonction.R
\name{gen_paragraphe}
\alias{gen_paragraphe}
\title{gen_paragraphe}
\usage{
gen_paragraphe(liste, nb = 25)
}
\arguments{
\item{liste}{the words to use}

\item{nb}{number of level}
}
\description{
generate level
}
\examples{
require(ztype)
require(magrittr)
c("dplyr","ggplot2","lubridate") \%>\% gen_liste() \%>\%
gen_paragraphe(10) \%>\% cat()
}

