% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction.R
\name{gen_set_of_levels}
\alias{gen_set_of_levels}
\title{gen_set_of_levels}
\usage{
gen_set_of_levels(set_of_words, nb = 25)
}
\arguments{
\item{set_of_words}{a sorted vector of the collection of words to use}

\item{nb}{the number of levels to generate}
}
\description{
generate a set of levels with increasing difficulty
}
\examples{
require(ztype)
require(magrittr)
c("dplyr","ggplot2","lubridate") \%>\% gen_set_of_words() \%>\%
gen_set_of_levels(10) \%>\% cat()
}

