% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction.R
\name{gen_set_of_words}
\alias{gen_set_of_words}
\title{gen_set_of_words}
\usage{
gen_set_of_words(packages)
}
\arguments{
\item{packages}{character vector. package(s) to parse and extract function names from}
}
\value{
a character vector
}
\description{
extracts all function names from a given list of packages
}
\examples{
\dontrun{
require(ztype)
require(magrittr)
c("dplyr","ggplot2","lubridate") \%>\% gen_set_of_words()
}
}

