% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction.R
\name{level}
\alias{level}
\title{level}
\usage{
level(set_of_words, quantity, difficulte)
}
\arguments{
\item{set_of_words}{a sorted vector of the collection of words to use}

\item{quantity}{an integer the number of words to pick in}

\item{difficulte}{an integer reflecting the difficulty level}
}
\description{
generates a collection of words to build a ZType game level
}
\examples{
require(ztype)
require(magrittr)
c("dplyr","ggplot2","lubridate") \%>\% gen_set_of_words() \%>\% level(10,50)
}

