% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fonction.R
\name{niveau}
\alias{niveau}
\title{niveau}
\usage{
niveau(liste, quantite, difficulte)
}
\arguments{
\item{liste}{the words to use}

\item{quantite}{number of word}

\item{difficulte}{difficulty}
}
\description{
generate a level
}
\examples{
require(ztype)
require(magrittr)
c("dplyr","ggplot2","lubridate") \%>\% gen_liste() \%>\% niveau(10,50)
}

