% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_package.R, R/fonction.R
\docType{package}
\encoding{UTF-8}
\name{ztype}
\alias{ztype}
\alias{ztype-package}
\title{ztype}
\usage{
ztype(packages = c("dplyr", "ggplot2", "lubridate"), nb = 25)
}
\arguments{
\item{packages}{vector with installed package names}

\item{nb}{number of level}
}
\description{
launch personnalized zty.pe game

launch a ztype game
}
\examples{

\dontrun{
require(ztype)
require(magrittr)
ztype() # dplyr, ggplot2 and lubridate
c("lubridate") \%>\% ztype()
}




\dontrun{
require(ztype)
require(magrittr)
ztype()# dplyr, ggplot2 and lubridate
c("lubridate") \%>\% ztype()
}

}

