% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction.R
\name{gen_game}
\alias{gen_game}
\title{gen_game}
\usage{
gen_game(set_of_levels, open = FALSE)
}
\arguments{
\item{set_of_levels}{the set of levels to use}

\item{open}{booleen open browser}
}
\description{
generate a ZType game by puhsing a set of levels on the website.
}
\examples{
\dontrun{
require(ztype)
require(magrittr)
c("dplyr","ggplot2","lubridate") \%>\% gen_set_of_words() \%>\%
gen_set_of_levels(10) \%>\%
gen_game() \%>\% browseURL()
}

}

