% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_package.R, R/fonction.R
\docType{package}
\encoding{UTF-8}
\name{ztype}
\alias{ztype}
\alias{ztype-package}
\title{ztype}
\usage{
ztype(packages = c("dplyr", "ggplot2", "lubridate"), nb = 25)
}
\arguments{
\item{packages}{a vector containing installed packages from which extract function names}

\item{nb}{number of levels to design in the ZType game}
}
\description{
How fast can you type R functions on your keyboard? Find out by running a 'zty.pe' game: export R functions as instructions to type to destroy opponents vessels.

launch a ZType game using function names of R packages
}
\examples{

\dontrun{
require(ztype)
require(magrittr)
ztype() # dplyr, ggplot2 and lubridate
c("lubridate") \%>\% ztype()
}




\dontrun{
require(ztype)
require(magrittr)
ztype()# dplyr, ggplot2 and lubridate
c("lubridate") \%>\% ztype()
}

}

